+++ scripts/openai/src/zap_openai/config.py
@@ -1,8 +1,11 @@
 from environs import Env
+from importlib.metadata import version
 
 env = Env()
 env.read_env()
 
-OPENAI_API_KEY = env.str("OPENAI_API_KEY")
-OPENAI_ORGANISATION_ID = env.str("OPENAI_ORGANISATION_ID")
-OPENAI_PROJECT_ID = env.str("OPENAI_PROJECT_ID")
+VERSION = version("zap-openai")
+
+OPENAI_API_KEY = env.str("APP_OPENAI_API_KEY")
+OPENAI_ORGANISATION_ID = env.str("APP_OPENAI_ORGANISATION_ID")
+OPENAI_PROJECT_ID = env.str("APP_OPENAI_PROJECT_ID")